Update packages for repository related objects
