+++ liveactivities/src/main/scala/com/gu/liveactivities/models/LiveActivities.scala
@@ -1,0 +1,43 @@
+package com.gu.liveactivities.models
+
+import play.api.libs.json._
+
+sealed trait LiveActivityData
+
+case class FootballLiveActivity(
+		homeTeam: String,
+		awayTeam: String,
+		articleUrl: String
+) extends LiveActivityData
+
+object FootballLiveActivity {
+	implicit val format: OFormat[FootballLiveActivity] =
+		Json.format[FootballLiveActivity]
+}
+
+object LiveActivityData {
+	implicit val format: OFormat[LiveActivityData] =
+		new OFormat[LiveActivityData] {
+			def writes(data: LiveActivityData): JsObject = data match {
+				case f: FootballLiveActivity =>
+					FootballLiveActivity.format.writes(f) + ("type" -> JsString(
+						"football"
+					))
+			}
+			def reads(json: JsValue): JsResult[LiveActivityData] =
+				(json \ "type").validate[String].flatMap {
+					case "football" => FootballLiveActivity.format.reads(json)
+					case other      => JsError(s"Unknown LiveActivityData type: $other")
+				}
+		}
+}
+
+case class LiveActivityMapping(
+		id: String,
+		channelId: String,
+		data: Option[LiveActivityData]
+)
+object LiveActivityMapping {
+	implicit val format: OFormat[LiveActivityMapping] =
+		Json.format[LiveActivityMapping]
+}
