+++ liveactivities/src/main/scala/com/gu/liveactivities/service/LiveActivityChannelRepository.scala
@@ -1,4 +1,4 @@
-package com.gu.liveactivities
+package com.gu.liveactivities.service
 
 import aws.AsyncDynamo
 import aws.DynamoJsonConversions.{fromAttributeMap, toAttributeMap}
@@ -11,49 +11,7 @@ import tracking.RepositoryError
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.jdk.CollectionConverters._
-
-// MODELS /////////////////////////////////////////////
-
-sealed trait LiveActivityData
-case class FootballLiveActivity(
-    homeTeam: String,
-    awayTeam: String,
-    articleUrl: String
-) extends LiveActivityData
-
-object FootballLiveActivity {
-  implicit val format: OFormat[FootballLiveActivity] =
-    Json.format[FootballLiveActivity]
-}
-
-object LiveActivityData {
-  implicit val format: OFormat[LiveActivityData] =
-    new OFormat[LiveActivityData] {
-      def writes(data: LiveActivityData): JsObject = data match {
-        case f: FootballLiveActivity =>
-          FootballLiveActivity.format.writes(f) + ("type" -> JsString(
-            "football"
-          ))
-      }
-      def reads(json: JsValue): JsResult[LiveActivityData] =
-        (json \ "type").validate[String].flatMap {
-          case "football" => FootballLiveActivity.format.reads(json)
-          case other      => JsError(s"Unknown LiveActivityData type: $other")
-        }
-    }
-}
-
-case class LiveActivityMapping(
-    id: String,
-    channelId: String,
-    data: Option[LiveActivityData]
-)
-object LiveActivityMapping {
-  implicit val format: OFormat[LiveActivityMapping] =
-    Json.format[LiveActivityMapping]
-}
-
-// REPOSITORY /////////////////////////////////////////////
+import com.gu.liveactivities.models.LiveActivityMapping
 
 trait ChannelMappingsRepository {
   def saveMapping(mapping: LiveActivityMapping): Future[RepositoryResult[Unit]]
@@ -68,6 +26,7 @@ class LiveActivityChannelRepository(client: AsyncDynamo, tableName: String)(
 ) extends ChannelMappingsRepository {
 
   private val logger: Logger = LoggerFactory.getLogger(this.getClass)
+  
   private val IdField = "id"
 
   override def saveMapping(
